refactor: DispatcherServletInitializer를 app 패키지로 이동
