+++ app/src/main/java/com/techcourse/Application.java
@@ -1,11 +1,10 @@
 package com.techcourse;
 
-import com.interface21.webmvc.servlet.mvc.TomcatStarter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.techcourse.support.TomcatStarter;
 import java.io.IOException;
 import java.util.stream.Stream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class Application {
