+++ app/src/main/java/com/techcourse/support/DispatcherServletInitializer.java
@@ -1,9 +1,10 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.techcourse.support;
 
+import com.interface21.web.WebApplicationInitializer;
+import com.interface21.webmvc.servlet.mvc.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.interface21.web.WebApplicationInitializer;
 
 /**
  * Base class for {@link WebApplicationInitializer}
